refactor: the withdrawal builder * moved the withdrawal builder to the via_verifier crate * Created new via_verifier_types crate and moved common types used by multiple crates
