+++ core/node/node_framework/src/implementations/layers/via_verifier/coordinator_api.rs
@@ -1,13 +1,10 @@
 use std::str::FromStr;
 
 use anyhow::Context;
-use via_btc_client::{
-    types::{BitcoinAddress, NodeAuth},
-    withdrawal_builder::WithdrawalBuilder,
-};
+use via_btc_client::types::{BitcoinAddress, NodeAuth};
 use via_btc_watch::BitcoinNetwork;
 use via_verifier_dal::{ConnectionPool, Verifier};
-use via_withdrawal_client::client::WithdrawalClient;
+use via_withdrawal_client::{client::WithdrawalClient, withdrawal_builder::WithdrawalBuilder};
 use zksync_config::{ViaBtcSenderConfig, ViaVerifierConfig};
 
 use crate::{
