+++ via_verifier/lib/via_musig2/Cargo.toml
@@ -14,6 +14,7 @@ categories.workspace = true
 rand.workspace = true
 hex.workspace = true
 via_btc_client.workspace = true
+via_verifier_types.workspace = true
 anyhow.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
@@ -33,7 +34,8 @@ uuid = { version = "1.3", features = ["v4"] }
 hyper = { version = "0.14", features = ["full"] }
 base64 = "0.21"
 
-
+[dev-dependencies]
+via_withdrawal_client.workspace = true
 
 [[example]]
 name = "key_generation_setup"
