+++ via_verifier/lib/via_verifier_types/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod transaction;
+pub mod withdrawal;
