+++ via_verifier/lib/via_verifier_types/src/withdrawal.rs
@@ -1,0 +1,7 @@
+use bitcoin::{Address, Amount};
+
+#[derive(Debug)]
+pub struct WithdrawalRequest {
+    pub address: Address,
+    pub amount: Amount,
+}
