+++ via_verifier/lib/via_withdrawal_client/Cargo.toml
@@ -24,11 +24,19 @@ via_btc_client.workspace = true
 via_da_client.workspace = true
 
 bitcoin = { version = "0.32.2", features = ["serde"] }
+serde.workspace = true
+via_verifier_types.workspace = true
+
+byteorder = "1.4"
+bincode = "1.3"
 
 [dev-dependencies]
 zksync_dal.workspace = true
 dotenv = "0.15"
 via_da_clients.workspace = true
+async-trait.workspace = true
+mockall = "0.13.0"
+bitcoincore-rpc = "0.19.0"
 
 [[example]]
 name = "withdraw"
