+++ via_verifier/lib/via_withdrawal_client/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod client;
 mod withdraw;
+pub mod withdrawal_builder;
